requirements.txt
python-telegram-bot==20.6
aiohttp==3.8.5
beautifulsoup4==4.12.2
schedule==1.2.0
python-dotenv==1.0.0
APScheduler==3.10.1